booking_system/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.2.1
Flask-Cors==3.0.10